requirements.txt
zai-sdk>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pymysql>=1.1.0
python-dotenv>=1.0.0
